app/requirements.txt
Flask==1.1.1
gunicorn==20
GitPython==3.1.7